requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain-core>=0.3.47,<1.0.0  # Loosened version constraint
langchain-groq>=0.3.1,<1.0.0   # Loosened version constraint
langgraph>=0.0.20,<1.0.0       # Loosened version constraint
langchain-openai>=0.0.1        # Added package
python-dotenv==1.0.1
ratelimit==2.2.1
pydantic>=2.7.4,<3.0.0  # Adjusted version constraint
typing-extensions==4.9.0
boto3==1.33.2
aioboto3==12.3.0
pytest==8.0.2
pytest-asyncio==0.23.5